Move fixtures to conftest.py to fix name re-use lints
